Updating Documentation
The Arduino Extension by Microsoft was missing when I followed the documentation, and OSX apps install to /Applications not /Application

The link to edit the documentation at https://docs.microsoft.com/en-us/samples/azure-samples/esp32-iot-devkit-get-started/sample/ is also broken.  You have to use the github link at the bottom to find the document to edit it.